cleanup, refactor & merge StreamingAdapter into BotFrameworkAdapter
